+++ src/main/java/power/keepeersofthestones/procedures/ClusterSharpIsGoneProcedure.java
@@ -60,7 +60,7 @@ public class ClusterSharpIsGoneProcedure {
 			}
 
 			private void run() {
-				world.setBlockState(new BlockPos((int) x, (int) y, (int) z), Blocks.AIR.getDefaultState(), 3);
+				world.setBlockState(new BlockPos(x, y, z), Blocks.AIR.getDefaultState(), 3);
 				MinecraftForge.EVENT_BUS.unregister(this);
 			}
 		}.start(world, (int) 400);
